全てのエラークラスをnako_errors.jsに置き、NakoErrorクラスを継承させる
