+++ test/plugin_system_test.js
@@ -1,6 +1,5 @@
 const assert = require('assert')
 const NakoCompiler = require('../src/nako3')
-const NakoRuntimeError = require('../src/nako_runtime_error')
 
 describe('plugin_system_test', () => {
   const nako = new NakoCompiler()
